Exclude parent pom 5.x from renovate updates
Projects configured from the archetype most often prefer to depend on
the last release of a recent LTS line.  The archetype currently requires
Jenkins 2.440.3 or newer.  When we switch to parent pom 5.x, we'll need
to require Jenkins 2.475 (weekly) or Jenkins 2.479.x (LTS).

I think it will be several months before we're ready to switch the
archetype to use plugin parent pom 5.x.  Until that time, this configures
renovate to only recommend versions less than 5.x